Update to VICE 2.4.20

+++ C64.emu/src/vice/monitor/mon_assemble65816.c
@@ -38,11 +38,11 @@
 #include "mon_util.h"
 #include "types.h"
 #include "uimon.h"
-
+#include "util.h"
 
 static int mon_assemble_instr(const char *opcode_name, asm_mode_addr_info_t operand)
 {
-    WORD operand_value = operand.param;
+    DWORD operand_value = operand.param;
     WORD operand_mode = operand.addr_mode;
     BYTE operand_extra_value = operand.addr_submode;
     BYTE i = 0, opcode = 0;
